This is the solution for the LogMeIn Cash Register Mayhem Problem.
2)     Cash Register Mayhem:
Design a web page that operates as a cash register. This should allow a
user to enter 1 to “n” items that contain the following fields:
-Item Name
-Item Price
-Quantity

This web page should allow the user to enter these and keep a running
total cost.  **Bonus points for adding the ability to calculate taxes!